perf: Wave 1+2 hot-path latency + bundle delta cleanup (5 fixes)
Closes 5 findings from /perf focused-delta scan (2026-04-25). Full per-scanner
output written to .perf-reports/perf-20260425-145039.md.

  1. PERF-DELTA-002 (HIGH 9/10) — account/delete: parallelize 3 independent
     ownership-marker writes (profile + sessions + device_tokens) and 3
     independent finalization writes (audit_log.insert + admin.updateUserById
     + supabase.auth.signOut). Each Promise.all collapses ~3-4 sequential
     network round-trips into one. Estimated savings: ~130ms per deletion.

  2. PERF-DELTA-003 (HIGH 8/10) — account/export: parallelize 4 sequential
     pre-flight id lookups (sessions, machines, webhooks, support_tickets)
     into a single Promise.all. None depend on each other; the only ordering
     constraint is that all 4 must finish before the main parallel block
     consumes the id arrays via .in() filters. Estimated savings: ~75ms.

  3. PERF-DELTA-005 (MEDIUM 9/10) — tier-enforcement.readPersonalTier:
     .single() → .maybeSingle(). PostgREST .single() throws PGRST116 when
     zero rows match, which is a legitimate case for new users with no
     subscription row yet. With .maybeSingle(), zero rows returns clean
     null without an error roundtrip. Eliminates Sentry noise + small CPU
     win on the gated-route hot path.

  4. PERF-DELTA-006 (MEDIUM 8/10) — webhooks/polar: avoid double JSON.parse
     of the request payload. handleTeamSubscriptionEvent now accepts the
     already-parsed object from the POST handler instead of re-parsing.
     Saves ~1-5ms per team webhook event.

  5. PERF-BUNDLE-001 (MEDIUM 9/10) — extract pure tier configuration from
     lib/polar.ts (which instantiates the Polar SDK at module scope) into
     lib/billing/tier-config.ts (zero SDK dependency). Update the two
     "use client" components on the public-facing /pricing page to import
     from the SDK-free module. Estimated savings: ~8-20 KB gzip on the
     public pricing page bundle (every visitor benefits, not just signed-in
     admins).

DEFERRED (documented residual / architectural):
  * PERF-DELTA-001 — cache polar_latest_order_id in subscriptions table at
    checkout time, eliminating polar.orders.list hop on refund. Requires
    new migration + webhook write integration. Saves 200-400ms per refund.
    Future enhancement; no PR opened.
  * PERF-DELTA-004 — orderIdOverride parameter on findRefundableOrderForSubscription.
    Already documented as future enhancement in the helper's WHY-comment.
  * PERF-DELTA-007 — parallelize generateSupportToken + headers in support
    actions. Saves ~5-10ms but the support flow already lives behind admin
    rate-limits; trivial gain not worth diff churn.
  * PERF-DELTA-008 — dynamic-import polar-refund inside issueRefundAction.
    Cold-start savings only on a low-traffic admin path; complexity > gain.
  * PERF-DELTA-011 — fire-and-forget audit_log.insert in webhook updated
    path. Saves ~25ms but tests check the await; not worth the test churn.
  * PERF-BUNDLE-002 — refactor support ticket detail page from full
    "use client" to server+island. Admin-only; UX would degrade (loading
    flash on the reply textarea). Future architectural decision.
  * server-only import on lib/polar.ts — wants `pnpm add server-only`.
    Defense-in-depth follow-up; the actual leak (TIERS-via-SDK pull) is
    already closed by the split.

Test plan:
  * 3140/3140 styrby-web tests pass. 12 route-test files updated to
    mock maybeSingle alongside single (chain-builder pattern).
  * 0 lint errors.
  * Build clean.
  * Type-check clean.
