+++ packages/styrby-web/src/app/api/account/export/route.ts
@@ -75,44 +75,29 @@ export async function POST(request: Request) {
     // WHY: Each table is independent, so parallel fetches are safe and faster.
     // Covers all 43 user-related tables for full GDPR Art. 15 / Art. 20 compliance.
     //
-    // Step 1: Fetch user's session IDs first so we can query session_messages.
-    // WHY: session_messages has no user_id column; it's linked via session_id.
-    const { data: userSessions } = await supabase
-      .from('sessions')
-      .select('id')
-      .eq('user_id', user.id)
-      .limit(10000);
+    // PERF-DELTA-003: Steps 1-4 fan-in id-lookups in parallel. None depends on
+    // any other; the only ordering constraint is that all four must finish
+    // before the main parallel block (which uses these id arrays in `.in()`
+    // filters). Saves ~75ms vs the prior sequential implementation.
+    const [
+      { data: userSessions },
+      { data: userMachines },
+      { data: userWebhooks },
+      { data: userTickets },
+    ] = await Promise.all([
+      // session_messages has no user_id column; it's linked via session_id.
+      supabase.from('sessions').select('id').eq('user_id', user.id).limit(10000),
+      // machine_keys has no direct user_id column — it is linked via machine_id.
+      supabase.from('machines').select('id').eq('user_id', user.id).limit(1000),
+      // webhook_deliveries has no direct user_id column — it is linked via webhook_id.
+      supabase.from('webhooks').select('id').eq('user_id', user.id).limit(1000),
+      // support_ticket_replies uses ticket_id (not user_id) as the ownership link.
+      supabase.from('support_tickets').select('id').eq('user_id', user.id).limit(1000),
+    ]);
 
     const sessionIds = (userSessions || []).map((s) => s.id);
-
-    // Step 2: Fetch machine IDs so we can query machine_keys via machine_id IN.
-    // WHY: machine_keys has no direct user_id column — it is linked via machine_id.
-    const { data: userMachines } = await supabase
-      .from('machines')
-      .select('id')
-      .eq('user_id', user.id)
-      .limit(1000);
-
     const machineIds = (userMachines || []).map((m) => m.id);
-
-    // Step 3: Fetch webhook IDs so we can query webhook_deliveries via webhook_id IN.
-    // WHY: webhook_deliveries has no direct user_id column — it is linked via webhook_id.
-    const { data: userWebhooks } = await supabase
-      .from('webhooks')
-      .select('id')
-      .eq('user_id', user.id)
-      .limit(1000);
-
     const webhookIds = (userWebhooks || []).map((w) => w.id);
-
-    // Step 4: Fetch support ticket IDs so we can query replies via ticket_id IN.
-    // WHY: support_ticket_replies uses ticket_id (not user_id) as the ownership link.
-    const { data: userTickets } = await supabase
-      .from('support_tickets')
-      .select('id')
-      .eq('user_id', user.id)
-      .limit(1000);
-
     const ticketIds = (userTickets || []).map((t) => t.id);
 
     // Step 5: Fetch all tables in parallel with limits to prevent OOM.
