+++ packages/styrby-web/src/app/api/bookmarks/__tests__/route.test.ts
@@ -30,6 +30,7 @@ function createChainMock() {
   }
 
   chain['single'] = vi.fn().mockResolvedValue(result);
+  chain['maybeSingle'] = vi.fn().mockResolvedValue(result);
   chain['then'] = (resolve: (v: unknown) => void) => resolve(result);
 
   return chain;
