+++ packages/styrby-web/src/app/api/invitations/__tests__/accept.test.ts
@@ -38,6 +38,7 @@ function createChainMock(result: { data?: unknown; error?: unknown }) {
     chain[method] = vi.fn().mockReturnValue(chain);
   }
   chain['single'] = vi.fn().mockResolvedValue(result);
+  chain['maybeSingle'] = vi.fn().mockResolvedValue(result);
   chain['then'] = (resolve: (v: unknown) => void) => resolve(result);
   return chain;
 }
