+++ packages/styrby-web/src/app/api/keys/__tests__/route.test.ts
@@ -44,6 +44,7 @@ function createChainMock() {
     chain[method] = vi.fn().mockReturnValue(chain);
   }
   chain['single'] = vi.fn().mockResolvedValue(result);
+  chain['maybeSingle'] = vi.fn().mockResolvedValue(result);
   chain['then'] = (resolve: (v: unknown) => void) => resolve(result);
   return chain;
 }
