+++ packages/styrby-web/src/app/api/teams/[id]/__tests__/route.test.ts
@@ -41,6 +41,7 @@ function createChainMock() {
   }
 
   chain['single'] = vi.fn().mockResolvedValue(result);
+  chain['maybeSingle'] = vi.fn().mockResolvedValue(result);
   chain['then'] = (resolve: (v: unknown) => void) => resolve(result);
 
   return chain;
