+++ packages/styrby-web/src/app/api/teams/[id]/members/__tests__/route.test.ts
@@ -42,6 +42,7 @@ function createChainMock() {
   }
 
   chain['single'] = vi.fn().mockResolvedValue(result);
+  chain['maybeSingle'] = vi.fn().mockResolvedValue(result);
   chain['then'] = (resolve: (v: unknown) => void) => resolve(result);
 
   return chain;
