+++ packages/styrby-web/src/app/api/teams/[id]/sso/__tests__/route.test.ts
@@ -46,6 +46,7 @@ function createChainMock(overrides?: { data?: unknown; error?: unknown; count?: 
     chain[m] = vi.fn().mockReturnValue(chain);
   }
   chain['single'] = vi.fn().mockResolvedValue(result);
+  chain['maybeSingle'] = vi.fn().mockResolvedValue(result);
   chain['then'] = (resolve: (v: unknown) => void) => resolve(result);
 
   return chain;
