+++ packages/styrby-web/src/app/api/webhooks/polar/route.ts
@@ -334,7 +334,13 @@ async function handleTeamSubscriptionEvent(
     canceled_at?: string;
     [key: string]: unknown;
   },
-  rawPayload: string
+  rawPayload: string,
+  // PERF-DELTA-006: rawParsed accepted from caller to avoid double-parse.
+  // The caller (POST handler) already parses once at line ~842 for routing
+  // decisions; passing the parsed object through saves a second JSON.parse
+  // on every team-subscription event (1-5ms per call on large payloads).
+  // rawPayload is still required separately for sha256Hex(rawPayload) below.
+  rawParsed: Record<string, unknown>
 ): Promise<Response> {
   // WHY admin client here and not before: createAdminClient() uses
   // SUPABASE_SERVICE_ROLE_KEY which bypasses RLS. We only call this after
@@ -352,9 +358,7 @@ async function handleTeamSubscriptionEvent(
   // key is the webhook event UUID (top-level `id` field), not the subscription ID.
   // The subscription ID can appear in multiple distinct events (created, updated,
   // canceled) — those must each be processed exactly once as separate events.
-  // We parse it safely from the raw payload to avoid trusting the already-cast
-  // `event` shape for a security-critical dedup key.
-  const rawParsed = JSON.parse(rawPayload) as Record<string, unknown>;
+  // We use the rawParsed object passed by the caller (no second parse here).
   const eventIdResult = PolarEventIdSchema.safeParse(rawParsed);
   if (!eventIdResult.success) {
     console.error('Team event missing top-level id field — cannot enforce idempotency');
@@ -864,7 +868,8 @@ export async function POST(request: Request) {
       event.type,
       teamId,
       teamDataResult.data,
-      payload
+      payload,
+      rawParsed,
     );
   }
